+++ build-tools/packages/build-tools/src/common/typeTests.ts
@@ -4,40 +4,7 @@
  */
 
 import path from "node:path";
-import { Project } from "ts-morph";
-import type { Package } from "../common/npmPackage";
-
-let shouldLog = false;
-export function enableLogging(enable: boolean) {
-	shouldLog = enable;
-}
-
-export function log(output: any) {
-	if (shouldLog) {
-		console.log(output);
-	}
-}
-
-/**
- * This uses the bit shifts instead of incrementing because it allows us to OR the
- * results of multiple checks together to get the largest breaking increment at the
- * end without needing to do any max(x,y) checks
- */
-export enum BreakingIncrement {
-	none = 0,
-	minor = 1,
-	major = (minor << 1) | minor,
-}
-
-export interface IValidator {
-	/**
-	 * Validate the internal state.  May mutate state and is only valid to call once
-	 * @param project - The Project which may be used to run a ts compilation task
-	 * @param pkgDir - The dir for the Project which may be used to create temporary
-	 *      source files
-	 */
-	validate(project: Project, pkgDir: string): BreakingIncrement;
-}
+import type { Package } from "./npmPackage";
 
 /**
  * Given a package, returns the name that should be used for the previous version of the package to generate type tests.
