+++ build-tools/packages/build-tools/src/fluidBuild/tasks/leaf/miscTasks.ts
@@ -8,7 +8,7 @@ import * as path from "path";
 import { readdir, stat } from "fs/promises";
 import picomatch from "picomatch";
 import { globFn, readFileAsync, statAsync, toPosixPath, unquote } from "../../../common/utils";
-import { getTypeTestPreviousPackageDetails } from "../../../typeValidator/validatorUtils";
+import { getTypeTestPreviousPackageDetails } from "../../../common/typeTests";
 import { BuildPackage } from "../../buildGraph";
 import { LeafTask, LeafWithFileStatDoneFileTask } from "./leafTask";
